Feat : Update Blog Model

+++ RestClientExample.RestClient/Models/BlogModel.cs
@@ -6,7 +6,7 @@
 public class BlogModel
 {
 	[Key]
-	public int BlogId {  get; set; }
+	public long BlogId {  get; set; }
 	public string BlogTitle { get; set; } = null!;
 	public string BlogAuthor { get; set; } = null!;
 	public string BlogContent { get; set; } = null!;
